Revert "Auth"
