+++ backend/package.json
@@ -7,7 +7,6 @@
     "fix": "prettier --write . && eslint --fix --ext .js ."
   },
   "dependencies": {
-    "argon2": "^0.30.3",
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
