--- backend/src/auth.js
@@ -1,28 +1,0 @@
-// hashing passwords
-const argon2 = require("argon2");
-
-const hashingOptions = {
-  type: argon2.argon2id,
-  memoryCost: 2 ** 16,
-  timeCost: 5,
-  parallelism: 1,
-};
-
-const hashPassword = (req, res, next) => {
-  argon2
-    .hash(req.body.pwd, hashingOptions)
-    .then((hashedPassword) => {
-      req.body.hashedPassword = hashedPassword;
-      delete req.body.pwd;
-
-      next();
-    })
-    .catch((err) => {
-      console.error(err);
-      res.sendStatus(500);
-    });
-};
-
-module.exports = {
-  hashPassword,
-};
