+++ backend/src/controllers/compoRecipeControllers.js
@@ -28,6 +28,22 @@ const read = (req, res) => {
     });
 };
 
+const getDetails = (req, res) => {
+  // get specific recipe details to show in resume page
+  models.compo_recipe
+    .get(req.params.id)
+    .then(([rows]) => {
+      if (rows[0] == null) {
+        res.sendStatus(404);
+      } else {
+        res.send(rows);
+      }
+    })
+    .catch((err) => {
+      console.error(err);
+      res.sendStatus(500);
+    });
+};
 const edit = (req, res) => {
   const compoRecipe = req.body;
 
@@ -85,7 +101,7 @@ const destroy = (req, res) => {
 module.exports = {
   browse,
   read,
-
+  getDetails,
   edit,
   add,
   destroy,
