+++ backend/src/controllers/userControllers.js
@@ -80,27 +80,10 @@ const destroy = (req, res) => {
     });
 };
 
-const getDetails = (req, res) => {
-  // get specific recipe details to show in resume page
-  models.compo_recipe
-    .get(req.params.id)
-    .then(([rows]) => {
-      if (rows[0] == null) {
-        res.sendStatus(404);
-      } else {
-        res.send(rows);
-      }
-    })
-    .catch((err) => {
-      console.error(err);
-      res.sendStatus(500);
-    });
-};
 module.exports = {
   browse,
   read,
   edit,
   add,
   destroy,
-  getDetails,
 };
