+++ backend/src/models/userManager.js
@@ -42,22 +42,6 @@ class UserManager extends AbstractManager {
       [account_id]
     );
   }
-
-  get(account_id) {
-    // this rrequest will be displayed on the resume page
-    return this.database.query(
-      `
-      SELECT cr.id, cr.percentage, cr.user_account_ID, u.firstname, u.lastname, u.address, u.birthdate, wb.bottle_name, r.recipe_name
-      FROM ${this.table} cr
-      JOIN user u ON cr.user_id = u.id
-      JOIN winebottle wb ON cr.wineBottle_id = wb.id
-      JOIN relation_recipe rr ON cr.id = rr.compoRecipe_id
-      JOIN recipe r ON rr.recipe_id = r.id
-      `,
-
-      [account_id]
-    );
-  }
 }
 
 module.exports = UserManager;
