+++ frontend/package.json
@@ -9,7 +9,6 @@
   "dependencies": {
     "@vitejs/plugin-react": "^3.0.1",
     "axios": "^1.4.0",
-    "gsap": "^3.12.2",
     "localforage": "^1.10.0",
     "match-sorter": "^6.3.1",
     "prop-types": "^15.8.1",
