+++ frontend/src/components/FormInfoPerso/FormInfoPerso.jsx
@@ -25,13 +25,12 @@ function FormInfoPerso() {
       });
   };
 
-
   // navigation
   const navigate = useNavigate();
 
-  // const handleClickNext = () => {
-  //   navigate("/FicheDegustation");
-  // };
+  const handleClickNext = () => {
+    navigate("/FicheDegustation");
+  };
 
   // stock mail & pwd du form pour post
   const [formAuthentification, setFormAuthentifiation] = useState({
@@ -164,10 +163,6 @@ function FormInfoPerso() {
         .then()
         .catch((err) => {
           console.error(err);
-        })
-        .finally(() => {
-          // Navigate to "/FicheDegustation"
-          navigate("/FicheDegustation");
         });
     }
   };
@@ -276,7 +271,7 @@ function FormInfoPerso() {
               Mot de Passe:
               <input
                 className="form-inscription-idendity-input-class"
-                type="password"
+                type="text"
                 name="pwd"
                 placeholder="........"
                 value={formAuthentification.pwd}
@@ -380,7 +375,7 @@ function FormInfoPerso() {
                   name="other-taste"
                   rows="5"
                   cols="33"
-                  placeholder="Précisez..."
+                  defaultValue="Précisez..."
                   value={formPreference.other}
                   onChange={handleOtherChange}
                 />
@@ -399,6 +394,10 @@ function FormInfoPerso() {
           </div>
         </div>
       </form>
+      <button type="button" onClick={handleClickNext}>
+        {" "}
+        Suivant
+      </button>
     </div>
   );
 }
