+++ frontend/src/components/FormInfoPerso/FormInfoPerso.scss
@@ -253,7 +253,6 @@
   font-family: $font_title_small;
   letter-spacing: 3px;
   position: inherit;
-  bottom: 5vh;
   @include tablet-paysage {
     width: 200px;
     position: absolute;
