+++ frontend/src/pages/AtelierCreation/AtelierCreation.jsx
@@ -1,5 +1,5 @@
-import Apercu from "../../components/Apercu/Apercu";
 import Bouteille from "../../components/Bouteille/Bouteille";
+import Apercu from "../../components/Apercu/Apercu";
 import "./AtelierCreation.scss";
 
 function AtelierCreation() {
