+++ frontend/src/pages/Resume/Resume.jsx
@@ -1,19 +1,16 @@
 import "./Resume.scss";
-import { useState, useEffect, useContext } from "react";
+import { useState, useEffect } from "react";
 import axios from "axios";
 import { useNavigate } from "react-router-dom";
 import bottle from "../../assets/images/bootle_360.png";
 import diploma from "../../assets/images/diplome.png";
 import fleche from "../../assets/images/fleche_360.png";
-import UserContext from "../../contexts/UserContext";
-import CommandeModal from "../../components/commandeModal/CommandeModal";
+// import UserContext from "../../contexts/UserContext";
 
 function Resume() {
   // useContext
-  const [user] = useContext(UserContext); // account_id of current user from inscription page, you can use it for update database
-
+  // const { user } = useContext(UserContext); // account_id of current user from inscription page, you can use it for update database
   // console.log(`resume account_id: ${user} `);
-  const [modalOpen, setModalOpen] = useState(false);
 
   const [fullName, setFullName] = useState("");
   const [emailAddress, setEmailAddress] = useState("");
@@ -21,14 +18,11 @@ function Resume() {
   const [bottleData, setBottleData] = useState([]);
   const [recipeName, setRecipeName] = useState("");
   const [quantity, setQuantity] = useState();
-
   const navigate = useNavigate();
-
+  const compoRecipeId = 1;
   useEffect(() => {
     axios
-      .get(`${import.meta.env.VITE_BACKEND_URL}/user`, {
-        params: { account_id: `${user}` },
-      })
+      .get(`${import.meta.env.VITE_BACKEND_URL}/compo_recipe/:${compoRecipeId}`)
       .then((response) => {
         const {
           firstname,
@@ -37,7 +31,7 @@ function Resume() {
           address,
           birthdate,
           recipe_name,
-        } = response.data[0];
+        } = response.data[1];
         setFullName(`${firstname} ${lastname}`);
         setEmailAddress(`${address}`);
         setBirthDay(`${birthdate}`);
@@ -52,10 +46,6 @@ function Resume() {
 
   const previousPage = () => {
     navigate("/livredor");
-  };
-
-  const openModal = () => {
-    setModalOpen(true);
   };
   return (
     <div className="resume_wrapper">
@@ -91,11 +81,9 @@ function Resume() {
             className="button-submit-identity"
             id="commande_btn"
             type="submit"
-            onClick={openModal}
           >
             Commander
           </button>
-          {modalOpen && <CommandeModal setOpenModal={setModalOpen} />}
         </div>
         <div className="bottle_diploma">
           <img id="bottle_resume" src={bottle} alt="" />
